Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    int r,pie = 3.14;
+    float r,pie = 3.14;
     printf("Area: ");
     scanf("%d", &r);
     float area = pie*(r*r);
